initial fallback setup complete
