+++ src/BugsnagUnity/Delivery.cs
@@ -14,6 +14,11 @@ namespace BugsnagUnity
     {
 
         private Configuration _configuration;
+
+        private CacheManager _cacheManager;
+
+        private PayloadManager _payloadManager;
+
         private object _callbackLock { get; } = new object();
 
         private static List<string> _finishedCacheDeliveries = new List<string>();
@@ -21,9 +26,11 @@ namespace BugsnagUnity
         private bool _cacheDeliveryInProcess;
 
 
-        internal Delivery(Configuration configuration)
+        internal Delivery(Configuration configuration, CacheManager cacheManager, PayloadManager payloadManager)
         {
             _configuration = configuration;
+            _cacheManager = cacheManager;
+            _payloadManager = payloadManager;
         }
 
         // Run any on send error callbacks if it's an event, serialise the payload and add it to the sending queue
@@ -100,12 +107,12 @@ namespace BugsnagUnity
                 if (code == 200 || code == 202)
                 {
                     // success!
-                    FileManager.PayloadSendSuccess(payload);
+                    _payloadManager.PayloadSendSuccess(payload);
                 }
                 else if (req.isNetworkError || code == 0 || code == 408 || code == 429 || code >= 500)
                 {
                     // sending failed with no network or retryable error, cache payload to disk
-                    FileManager.SendPayloadFailed(payload);
+                    _payloadManager.SendPayloadFailed(payload);
                 }
                 _finishedCacheDeliveries.Add(payload.Id);
             }
@@ -131,14 +138,12 @@ namespace BugsnagUnity
 
         private IEnumerator DeliverCachedPayloads()
         {
-            foreach (var cachedPayloadPath in FileManager.GetCachedPayloadPaths())
+            var payload = _cacheManager.GetNextCachedPayload();
+            while (payload != null)
             {
-                var payload = FileManager.GetPayloadFromCachePath(cachedPayloadPath);
-                if (payload != null)
-                {
-                    Deliver(payload);
-                    yield return new WaitUntil(() => CachedPayloadProcessed(payload.Id));
-                }
+                Deliver(payload);
+                yield return new WaitUntil(() => CachedPayloadProcessed(payload.Id));
+                payload = _cacheManager.GetNextCachedPayload();
             }
             _cacheDeliveryInProcess = false;
         }
